Add run sub command
