+++ oremda/cli/run.py
@@ -1,9 +1,7 @@
 #!/usr/bin/env python
 import json
 import os
-from typing import Dict
-
-import matplotlib.pyplot as plt
+import click
 
 from oremda.clients import Client as ContainerClient
 from oremda.plasma_client import PlasmaClient
@@ -15,119 +13,73 @@ from oremda.constants import (
 )
 from oremda.utils.plasma import start_plasma_store
 from oremda.typing import ContainerType, DisplayType, IdType, Port
-from oremda.display import DisplayHandle, NoopDisplayHandle
+from oremda.display import display_factory
 import oremda.pipeline
 
 
-class MatplotlibDisplayHandle(DisplayHandle):
-    def __init__(self, id: IdType):
-        super().__init__(id, DisplayType.OneD)
-        self.inputs: Dict[IdType, Port] = {}
 
-    def add(self, sourceId: IdType, input: Port):
-        self.inputs[sourceId] = input
-        self.render()
+@click.command('run', short_help='oremda cli runner', help='Run an oremda pipeline on the command line.')
+@click.argument('pipeline', type=click.Path(exists=True, dir_okay=False))
+def main(pipeline):
+    if "SINGULARITY_BIND" in os.environ:
+        # This means we are running singularity
+        container_type = ContainerType.Singularity
 
-    def remove(self, sourceId: IdType):
-        if id in self.inputs:
-            del self.inputs[sourceId]
-        self.render()
-
-    def clear(self):
-        self.inputs = {}
-        self.render()
-
-    def render(self):
-        x_label = self.parameters.get("xLabel", "x")
-        y_label = self.parameters.get("yLabel", "y")
-
-        fg, ax = plt.subplots(1, 1)
-
-        for port in self.inputs.values():
-            if port.data is None:
-                continue
-
-            x = port.data.data[0]
-            y = port.data.data[1]
-
-            label = None
-            if port.meta is not None:
-                label = port.meta.get("label")
-
-            ax.plot(x, y, label=label)
-
-        ax.legend()
-        ax.set(xlabel=x_label, ylabel=y_label)
-
-        fg.savefig(f"/data/{self.id}.png", dpi=fg.dpi)
-
-
-def display_factory(id: IdType, display_type: DisplayType) -> DisplayHandle:
-    if display_type == DisplayType.OneD:
-        return MatplotlibDisplayHandle(id)
+        # Remove this so we don't repeat the parent container's bind mounting
+        del os.environ["SINGULARITY_BIND"]
     else:
-        return NoopDisplayHandle(id, display_type)
+        container_type = ContainerType.Docker
 
-
-if "SINGULARITY_BIND" in os.environ:
-    # This means we are running singularity
-    container_type = ContainerType.Singularity
-
-    # Remove this so we don't repeat the parent container's bind mounting
-    del os.environ["SINGULARITY_BIND"]
-else:
-    container_type = ContainerType.Docker
-
-plasma_kwargs = {
-    "memory": 50_000_000,
-    "socket_path": DEFAULT_PLASMA_SOCKET_PATH,
-}
-
-with start_plasma_store(**plasma_kwargs):
-    plasma_client = PlasmaClient(DEFAULT_PLASMA_SOCKET_PATH)
-    container_client = ContainerClient(container_type)
-
-    if container_type == ContainerType.Singularity:
-        container_client.images_dir = "/images"
-
-    registry = Registry(plasma_client, container_client)
-
-    run_kwargs = {
-        "detach": True,
-        "working_dir": DEFAULT_DATA_DIR,
+    plasma_kwargs = {
+        "memory": 50_000_000,
+        "socket_path": DEFAULT_PLASMA_SOCKET_PATH,
     }
 
-    if container_type == ContainerType.Singularity:
-        # The mounts in singularity child containers refer to the
-        # parent container's directories.
-        volumes = {
-            DEFAULT_OREMDA_VAR_DIR: {"bind": DEFAULT_OREMDA_VAR_DIR},
-            DEFAULT_DATA_DIR: {"bind": DEFAULT_DATA_DIR},
-            "/oremda": {"bind": "/oremda"},
+    with start_plasma_store(**plasma_kwargs):
+        plasma_client = PlasmaClient(DEFAULT_PLASMA_SOCKET_PATH)
+        container_client = ContainerClient(container_type)
+
+        if container_type == ContainerType.Singularity:
+            container_client.images_dir = "/images"
+
+        registry = Registry(plasma_client, container_client)
+
+        run_kwargs = {
+            "detach": True,
+            "working_dir": DEFAULT_DATA_DIR,
         }
-        run_kwargs["volumes"] = volumes
-    if container_type == ContainerType.Docker:
-        # The mounts in docker "sibling" containers refer to the
-        # host directories.
-        self_container = container_client.self_container()
-        volumes = {
-            mount.source: {"bind": mount.destination} for mount in self_container.mounts
-        }
-        run_kwargs["volumes"] = volumes
 
-        # Get the child containers to share IPC with the parent
-        run_kwargs["ipc_mode"] = f"container:{self_container.id}"
+        if container_type == ContainerType.Singularity:
+            # The mounts in singularity child containers refer to the
+            # parent container's directories.
+            volumes = {
+                DEFAULT_OREMDA_VAR_DIR: {"bind": DEFAULT_OREMDA_VAR_DIR},
+                DEFAULT_DATA_DIR: {"bind": DEFAULT_DATA_DIR},
+                "/oremda": {"bind": "/oremda"},
+            }
+            run_kwargs["volumes"] = volumes
+        if container_type == ContainerType.Docker:
+            # The mounts in docker "sibling" containers refer to the
+            # host directories.
+            self_container = container_client.self_container()
+            volumes = {
+                mount.source: {"bind": mount.destination} for mount in self_container.mounts
+            }
+            run_kwargs["volumes"] = volumes
 
-    registry.run_kwargs = run_kwargs
+            # Get the child containers to share IPC with the parent
+            run_kwargs["ipc_mode"] = f"container:{self_container.id}"
 
-    with open("/runner/pipeline.json") as f:
-        pipeline_obj = json.load(f)
+        registry.run_kwargs = run_kwargs
 
-    pipeline = oremda.pipeline.deserialize_pipeline(
-        pipeline_obj, plasma_client, registry, display_factory
-    )
+        with open(pipeline) as f:
+            pipeline_obj = json.load(f)
 
-    try:
-        pipeline.run()
-    finally:
-        registry.release()
+        pipeline = oremda.pipeline.deserialize_pipeline(
+            pipeline_obj, plasma_client, registry, display_factory
+        )
+
+        try:
+            pipeline.run()
+        finally:
+            registry.release()
